move the mainteance banner back to the landing page
